Adding user role access criterias to APIs

+++ kontizePlugin/src/main/java/no/haagensoftware/kontize/handler/adminKeysHandler.java
@@ -9,6 +9,7 @@ import no.haagensoftware.kontize.db.LevelDbEnv;
 import no.haagensoftware.kontize.models.AdminKey;
 import no.haagensoftware.kontize.models.AdminKeyList;
 import no.haagensoftware.kontize.models.AdminKeyObject;
+import no.haagensoftware.kontize.models.AuthenticationResult;
 import org.apache.log4j.Logger;
 
 import java.util.List;
@@ -18,8 +19,10 @@ import java.util.List;
  */
 public class AdminKeysHandler extends ContenticeHandler {
     private static final Logger logger = Logger.getLogger(AdminKeysHandler.class.getName());
+    private AuthenticationContext authenticationContext;
 
     public AdminKeysHandler() {
+         authenticationContext = AuthenticationContext.getInstance();
     }
 
     @Override
@@ -28,16 +31,25 @@ public class AdminKeysHandler extends ContenticeHandler {
 
         String adminDataTypeId = getParameter("adminKey");
 
-        if (isGet(fullHttpRequest) && adminDataTypeId != null) {
+        String cookieUuidToken = getCookieValue(fullHttpRequest, "uuidToken");
+        AuthenticationResult cachedUserResult = null;
+        if (cookieUuidToken != null) {
+            cachedUserResult = authenticationContext.verifyUUidToken(cookieUuidToken);
+            logger.info("cachedUserResult: " + cachedUserResult);
+        }
+
+        boolean isAdmin = authenticationContext.isAdmin(cachedUserResult.getUuidToken(), cachedUserResult.getUserId());
+
+        if (isAdmin && isGet(fullHttpRequest) && adminDataTypeId != null) {
             //Get single admin data type
-        } else if (isGet(fullHttpRequest) && adminDataTypeId == null) {
+        } else if (isAdmin && isGet(fullHttpRequest) && adminDataTypeId == null) {
              List<AdminKey> adminKeys = LevelDbEnv.getInstance().getUserDao().getKeys();
 
             AdminKeyList adminKeyList = new AdminKeyList();
             adminKeyList.setAdminKeys(adminKeys);
 
             jsonReturn = new Gson().toJson(adminKeyList);
-        } else if (isPut(fullHttpRequest) && adminDataTypeId != null) {
+        } else if (isAdmin && isPut(fullHttpRequest) && adminDataTypeId != null) {
             String messageContent = getHttpMessageContent(fullHttpRequest);
             AdminKeyObject adminKeyObject = new Gson().fromJson(messageContent, AdminKeyObject.class);
             adminKeyObject.getAdminKey().setId(adminDataTypeId);
@@ -48,10 +60,14 @@ public class AdminKeysHandler extends ContenticeHandler {
             LevelDbEnv.getInstance().getUserDao().storeKey(adminKey.getId(), adminKey.getValue());
 
             jsonReturn = new Gson().toJson(adminKeyObject);
-        } else if (isDelete(fullHttpRequest) && adminDataTypeId != null) {
+        } else if (isAdmin && isDelete(fullHttpRequest) && adminDataTypeId != null) {
             LevelDbEnv.getInstance().getUserDao().deleteKey(adminDataTypeId);
         }
 
+        if (jsonReturn.length() == 0) {
+            AdminKeyList list = new AdminKeyList();
+            jsonReturn = new Gson().toJson(list);
+        }
         writeContentsToBuffer(channelHandlerContext, jsonReturn, "application/json; charset=UTF-8");
     }
